+++ packages/amplify-cli/src/index.ts
@@ -36,7 +36,6 @@ import { getAmplifyVersion } from './extensions/amplify-helpers/get-amplify-vers
 import { init as initErrorHandler, handleException, handleUnhandledRejection } from './amplify-exception-handler';
 
 export { UsageData } from './domain/amplify-usageData';
-export { getProjectSettings } from './context-manager';
 
 // Adjust defaultMaxListeners to make sure Inquirer will not fail under Windows because of the multiple subscriptions
 // https://github.com/SBoudrias/Inquirer.js/issues/887
@@ -258,4 +257,4 @@ export const executeAmplifyCommand = async (context: Context): Promise<void> => 
   }
 };
 
-// bump version to 12.13.0
+// bump version to 12.14.0