+++ scripts/cloud-e2e.sh
@@ -16,7 +16,14 @@ function cloudE2EBeta {
 }
 function cloudE2E {
     echo Running Prod E2E Test Suite
-    export TARGET_BRANCH=run-cb-e2e/$USER/$CURR_BRANCH
+    if [[ -n "$USER" ]]; then
+        export TARGET_BRANCH=run-cb-e2e/$USER/$CURR_BRANCH
+    elif [[ -n "$USERNAME" ]]; then
+        export TARGET_BRANCH=run-cb-e2e/$USERNAME/$CURR_BRANCH
+    else
+        echo "Error: Both USER and USERNAME variables are not set."
+        exit 1
+    fi
     git push $(git remote -v | grep aws-amplify/amplify-cli | head -n1 | awk '{print $1;}') $CURR_BRANCH:$TARGET_BRANCH --no-verify --force-with-lease
     triggerProjectBatch $E2E_ACCOUNT_PROD $E2E_ROLE_NAME "${E2E_PROFILE_NAME}Prod" $E2E_PROJECT_NAME $TARGET_BRANCH
 }