+++ packages/amplify-app/CHANGELOG.md
@@ -3,7 +3,7 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
-## [5.0.37-gen2-migration-test-alpha.0](https://github.com/aws-amplify/amplify-cli/compare/@aws-amplify/amplify-app@5.0.37-gen2-migrations-test.0...@aws-amplify/amplify-app@5.0.37-gen2-migration-test-alpha.0) (2024-09-26)
+## [5.0.38](https://github.com/aws-amplify/amplify-cli/compare/@aws-amplify/amplify-app@5.0.37...@aws-amplify/amplify-app@5.0.38) (2025-01-30)
 
 **Note:** Version bump only for package @aws-amplify/amplify-app
 
@@ -11,7 +11,7 @@ See [Conventional Commits](https://conventionalcommits.org) for commit guideline
 
 
 
-## [5.0.37-gen2-migrations-test.0](https://github.com/aws-amplify/amplify-cli/compare/@aws-amplify/amplify-app@5.0.36...@aws-amplify/amplify-app@5.0.37-gen2-migrations-test.0) (2024-09-23)
+## [5.0.37](https://github.com/aws-amplify/amplify-cli/compare/@aws-amplify/amplify-app@5.0.36...@aws-amplify/amplify-app@5.0.37) (2025-01-02)
 
 **Note:** Version bump only for package @aws-amplify/amplify-app