+++ packages/amplify-category-notifications/CHANGELOG.md
@@ -3,7 +3,7 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
-## [2.26.24-alpha.0](https://github.com/aws-amplify/amplify-cli/compare/@aws-amplify/amplify-category-notifications@2.26.24-gen2-migration-test-alpha.0...@aws-amplify/amplify-category-notifications@2.26.24-alpha.0) (2024-11-21)
+## [2.26.27](https://github.com/aws-amplify/amplify-cli/compare/@aws-amplify/amplify-category-notifications@2.26.26...@aws-amplify/amplify-category-notifications@2.26.27) (2025-01-30)
 
 **Note:** Version bump only for package @aws-amplify/amplify-category-notifications
 
@@ -11,7 +11,7 @@ See [Conventional Commits](https://conventionalcommits.org) for commit guideline
 
 
 
-## [2.26.24-gen2-migration-test-alpha.0](https://github.com/aws-amplify/amplify-cli/compare/@aws-amplify/amplify-category-notifications@2.26.24-gen2-migrations-test.0...@aws-amplify/amplify-category-notifications@2.26.24-gen2-migration-test-alpha.0) (2024-09-26)
+## [2.26.26](https://github.com/aws-amplify/amplify-cli/compare/@aws-amplify/amplify-category-notifications@2.26.25...@aws-amplify/amplify-category-notifications@2.26.26) (2025-01-16)
 
 **Note:** Version bump only for package @aws-amplify/amplify-category-notifications
 
@@ -19,7 +19,15 @@ See [Conventional Commits](https://conventionalcommits.org) for commit guideline
 
 
 
-## [2.26.24-gen2-migrations-test.0](https://github.com/aws-amplify/amplify-cli/compare/@aws-amplify/amplify-category-notifications@2.26.23...@aws-amplify/amplify-category-notifications@2.26.24-gen2-migrations-test.0) (2024-09-23)
+## [2.26.25](https://github.com/aws-amplify/amplify-cli/compare/@aws-amplify/amplify-category-notifications@2.26.24...@aws-amplify/amplify-category-notifications@2.26.25) (2024-11-18)
+
+**Note:** Version bump only for package @aws-amplify/amplify-category-notifications
+
+
+
+
+
+## [2.26.24](https://github.com/aws-amplify/amplify-cli/compare/@aws-amplify/amplify-category-notifications@2.26.23...@aws-amplify/amplify-category-notifications@2.26.24) (2024-10-05)
 
 **Note:** Version bump only for package @aws-amplify/amplify-category-notifications